femur_fracture_app/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.20.0
tkinter
torch>=1.12.0
torchvision>=0.13.0
matplotlib>=3.5.0